refactor(Notification): contract uplift + REST→direct calls
- Lifted 16 contract DTOs to dristi-common/contract/notification/
  (Notification, NotificationRequest/Response/ListResponse/SearchRequest,
  NotificationExists/ExistsRequest/ExistsResponse, NotificationCriteria,
  Order, Pagination, Id*/RequestInfoWrapper)
- 0 REST calls to convert (no intra-DRISTI candidates detected)
- Post-lift: removed stale internal.web.models.* wildcard import from
  NotificationApiController, NotificationService, and their tests
