+++ dristi-monolith/domain-case-lifecycle/src/main/java/org/pucar/dristi/caselifecycle/notification/internal/repository/NotificationRepository.java
@@ -3,10 +3,10 @@ package org.pucar.dristi.caselifecycle.notification.internal.repository;
 import lombok.extern.slf4j.Slf4j;
 import org.pucar.dristi.caselifecycle.notification.internal.repository.querybuilder.NotificationQueryBuilder;
 import org.pucar.dristi.caselifecycle.notification.internal.repository.rowmapper.NotificationRowMapper;
-import org.pucar.dristi.caselifecycle.notification.internal.web.models.Notification;
-import org.pucar.dristi.caselifecycle.notification.internal.web.models.NotificationCriteria;
-import org.pucar.dristi.caselifecycle.notification.internal.web.models.NotificationExists;
-import org.pucar.dristi.caselifecycle.notification.internal.web.models.Pagination;
+import org.pucar.dristi.common.contract.notification.Notification;
+import org.pucar.dristi.common.contract.notification.NotificationCriteria;
+import org.pucar.dristi.common.contract.notification.NotificationExists;
+import org.pucar.dristi.common.contract.notification.Pagination;
 import org.egov.tracer.model.CustomException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
