+++ dristi-monolith/domain-case-lifecycle/src/main/java/org/pucar/dristi/caselifecycle/notification/internal/repository/querybuilder/NotificationQueryBuilder.java
@@ -1,9 +1,9 @@
 package org.pucar.dristi.caselifecycle.notification.internal.repository.querybuilder;
 
 import lombok.extern.slf4j.Slf4j;
-import org.pucar.dristi.caselifecycle.notification.internal.web.models.NotificationCriteria;
-import org.pucar.dristi.caselifecycle.notification.internal.web.models.NotificationExists;
-import org.pucar.dristi.caselifecycle.notification.internal.web.models.Pagination;
+import org.pucar.dristi.common.contract.notification.NotificationCriteria;
+import org.pucar.dristi.common.contract.notification.NotificationExists;
+import org.pucar.dristi.common.contract.notification.Pagination;
 import org.egov.tracer.model.CustomException;
 import org.springframework.stereotype.Component;
