+++ dristi-monolith/domain-case-lifecycle/src/main/java/org/pucar/dristi/caselifecycle/notification/internal/service/WorkflowService.java
@@ -3,8 +3,8 @@ package org.pucar.dristi.caselifecycle.notification.internal.service;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.pucar.dristi.caselifecycle.notification.internal.config.Configuration;
 import org.pucar.dristi.common.repository.ServiceRequestRepository;
-import org.pucar.dristi.caselifecycle.notification.internal.web.models.Notification;
-import org.pucar.dristi.caselifecycle.notification.internal.web.models.NotificationRequest;
+import org.pucar.dristi.common.contract.notification.Notification;
+import org.pucar.dristi.common.contract.notification.NotificationRequest;
 import org.egov.common.contract.models.RequestInfoWrapper;
 import org.egov.common.contract.models.Workflow;
 import org.egov.common.contract.request.RequestInfo;
