+++ dristi-monolith/dristi-common/src/main/java/org/pucar/dristi/common/contract/notification/NotificationListResponse.java
@@ -1,4 +1,5 @@
-package org.pucar.dristi.caselifecycle.notification.internal.web.models;
+// HAND-CURATED — lifted by Phase 35 (contract-lift)
+package org.pucar.dristi.common.contract.notification;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import jakarta.validation.Valid;
@@ -13,7 +14,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 /**
- * NotificationExistsResponse
+ * NotificationListResponse
  */
 @Validated
 @jakarta.annotation.Generated(value = "org.egov.codegen.SpringBootCodegen", date = "2025-02-07T11:59:26.022967807+05:30[Asia/Kolkata]")
@@ -21,22 +22,29 @@ import java.util.List;
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
-public class NotificationExistsResponse {
+public class NotificationListResponse {
 
     @JsonProperty("ResponseInfo")
     @Valid
     private ResponseInfo responseInfo = null;
 
-    @JsonProperty("notificationList")
+    @JsonProperty("totalCount")
+    private Integer totalCount = null;
+
+    @JsonProperty("list")
     @Valid
-    private List<NotificationExists> notificationList = null;
+    private List<Notification> list = null;
+
+    @JsonProperty("pagination")
+    @Valid
+    private Pagination pagination = null;
 
 
-    public NotificationExistsResponse addNotificationListItem(NotificationExists notificationListItem) {
-        if (this.notificationList == null) {
-            this.notificationList = new ArrayList<>();
+    public NotificationListResponse addListItem(Notification listItem) {
+        if (this.list == null) {
+            this.list = new ArrayList<>();
         }
-        this.notificationList.add(notificationListItem);
+        this.list.add(listItem);
         return this;
     }
