+++ dristi-monolith/dristi-common/src/main/java/org/pucar/dristi/common/contract/notification/NotificationSearchRequest.java
@@ -1,4 +1,5 @@
-package org.pucar.dristi.caselifecycle.notification.internal.web.models;
+// HAND-CURATED — lifted by Phase 35 (contract-lift)
+package org.pucar.dristi.common.contract.notification;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import jakarta.validation.Valid;
@@ -10,7 +11,7 @@ import org.egov.common.contract.request.RequestInfo;
 import org.springframework.validation.annotation.Validated;
 
 /**
- * NotificationRequest
+ * NotificationSearchRequest
  */
 @Validated
 @jakarta.annotation.Generated(value = "org.egov.codegen.SpringBootCodegen", date = "2025-02-07T11:59:26.022967807+05:30[Asia/Kolkata]")
@@ -18,15 +19,19 @@ import org.springframework.validation.annotation.Validated;
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
-public class NotificationRequest {
+public class NotificationSearchRequest {
 
     @JsonProperty("RequestInfo")
     @Valid
     private RequestInfo requestInfo = null;
 
-    @JsonProperty("notification")
+    @JsonProperty("criteria")
     @Valid
-    private Notification notification = null;
+    private NotificationCriteria criteria = null;
+
+    @JsonProperty("pagination")
+    @Valid
+    private Pagination pagination = null;
 
 
 }
